# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Первый шаг API
  version: 1.0.0
  description: Same-origin API для MAX Mini App; demo mode uses an HttpOnly local session.
servers:
  - url: /
security:
  - maxInitData: []
paths:
...
  /api/max/webhook:
    post:
      security: []
      description: MAX Webhook endpoint; authenticated with X-Max-Bot-Api-Secret.
      responses:
        '200': {description: Update accepted}
        '403': {description: Invalid webhook secret}
...
  /api/profile:
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProfilePatch'}
      responses:
        '200': {description: Saved profile}
  /api/plans:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId]
              properties:
                eventId: {type: string}
                mode: {type: string, enum: [friend, solo]}
                when: {type: string, format: date-time, nullable: true}
                meeting: {type: string, maxLength: 240}
      responses:
        '201': {description: Created personal plan}
        '400': {description: Validation error}
  /api/plans/{id}:
    patch:
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PlanPatch'}
      responses:
        '200': {description: Updated plan}
        '403': {description: Only owner can edit}
...
components:
  securitySchemes:
    maxInitData:
      type: apiKey
      in: header
      name: X-Max-Init-Data
      description: MAX WebApp.initData; server validates HMAC with MAX bot token. Demo mode uses HttpOnly first_session instead.
  schemas:
    ProfilePatch:
      type: object
      properties:
        category: {type: string, enum: [all, animals, people]}
        barrier: {type: string, enum: [company, unknown, time]}
        reminders: {type: boolean}
    PlanPatch:
      type: object
      properties:
        when: {type: string, format: date-time, nullable: true}
        meeting: {type: string, maxLength: 240}
        confirmed: {type: boolean}
        checks: {type: array, items: {type: string, enum: [contact, route, bag]}}
        status: {type: string, enum: [done, cancelled]}
        reflection: {type: string, enum: [warm, okay, hard]}